requirements.txt
requests>=2.32.0
google-genai>=1.10.0
Flask>=3.1.0
playwright>=1.52.0
gunicorn>=23.0.0
